convert all kubernetes manifests to jsonnet
The manifests for the cert-manager application were in jsonnet and manifested
into cue files. If the jsonnet code imported secrets, then the converted cue
files would contain the sensitive information.

A workaround would be inject the sensitive information with an additional cue
file. This solution is ignored because the configuration for the cert-manager
application would feel scaterred.

Another workaround would be to omit the downstream.cue from the repository and
add a .gitignore for the file. This solution is also ignored because it is using
.gitignore to patch a design limitation.

As a result, I've decided to fully commit to jsonnet for the kubernetes
component of this repository.